Extract auth in a nicer way
